backend/requirements.txt
blinker==1.9.0
boto3==1.35.87
botocore==1.35.87
click==8.1.8
ecdsa==0.19.0
Flask==3.1.0
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
MarkupSafe==3.0.2
pyasn1==0.6.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-jose==3.3.0
rsa==4.9
s3transfer==0.10.4
six==1.17.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
python-dotenv==1.0.1
Flask-Migrate==4.0.7
